get events by category endpoind added to front
